[tests] Renamed selenium test file for consistency
